add fluxoperator-gen and setup for helper commands
by way of refactoring to make several functions public and separate from the reconciler we can generate separate helper clients. The first is fluxoperator-gen that helps to dump the raw yaml manifests for the flux operator to use in other contexts
