+++ config/crd/bases/flux-framework.org_miniclusters.yaml
@@ -241,6 +241,7 @@ spec:
                           items:
                             type: string
                           type: array
+                          x-kubernetes-list-type: atomic
                         privileged:
                           description: Privileged container
                           type: boolean
@@ -302,6 +303,7 @@ spec:
                               type: integer
                           type: object
                         type: array
+                        x-kubernetes-list-type: atomic
                       hostlist:
                         description: Hostlist is a custom hostlist for the broker.toml
                           that includes the local plus bursted cluster. This is typically
@@ -682,6 +684,7 @@ spec:
                           items:
                             type: string
                           type: array
+                          x-kubernetes-list-type: atomic
                         privileged:
                           description: Privileged container
                           type: boolean
