+++ docs/getting_started/index.md
@@ -9,5 +9,6 @@ any questions or issues, please [let us know](https://github.com/flux-framework/
 ```{toctree}
 :maxdepth: 2
 user-guide
+helper-clients
 custom-resource-definition
 ```
